backend/requirements.txt
flask>=3.0.0
flask-restful>=0.3.10
flask-cors>=4.0.0
web3>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
groq>=0.9.0
openai>=1.0.0
google-generativeai>=0.5.0
